requirements.txt
torch
pytest
numpy
tqdm
pillow
plyfile
viser
mlx
nanobind
setuptools